ミドルウェアを追加、ビルドエラーがあったので、そちらを対応。店舗登録・店舗編集画面、画像アップロード画面のpage.tsxをサーバーコンポーネントに変更
