+++ next.config.ts
@@ -15,13 +15,6 @@ const nextConfig: NextConfig = {
     serverActions: {
       bodySizeLimit: "6mb"
     }
-  },
-  webpack: (config, { isServer }) => {
-    if (!isServer) {
-      // 'firebase-admin' をクライアントバンドルから除外する
-      config.externals.push('firebase-admin')
-    }
-    return config
   }
 };
