+++ src/components/Store/StoreCloseActionsPanel.tsx
@@ -5,8 +5,8 @@ import { Box, Button } from "@mui/material";
 import { useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import { Dispatch, SetStateAction, useState } from "react";
-import { ConfirmDialog } from "./modals/ConfirmDialog";
-import { ResultDialog } from "./modals/ResultDialog";
+import { ConfirmDialog } from "../modals/ConfirmDialog";
+import { ResultDialog } from "../modals/ResultDialog";
 import { getDisplayStoreName } from "@/utils/storeUtils";
 import { useCloseStore } from "@/hooks/api/useStores";
 import { useAuthStore } from "@/lib/AuthStore";
