+++ src/middleware.ts
@@ -1,25 +1,35 @@
 import { NextRequest, NextResponse } from "next/server";
-import { verifySession } from "./lib/server/firebaseAdmin";
 
-export const runtime = 'nodejs'
+export const runtime = 'experimental-edge'
 
-// この関数は、リクエストがmatcheに指定されたパスと一致する場合に呼び出されます。
+// この関数は、リクエストがmatchesに指定されたパスと一致する場合に呼び出されます。
 export async function middleware(request: NextRequest) {
-    const user = await verifySession(request)
+    const session = request.cookies.get('session')?.value
 
-    if (!user) {
-        // ユーザーが認証されていない場合、ログインページにリダイレクトします。
-        // request.urlを使用して、リダイレクト後のログイン成功時に
-        // 元のページに戻れるよう、リダイレクトURLをクエリパラメータとして追加します。
-        const loginUrl = new URL(`/auth/login`, request.url)
+    if (!session) {
+        const loginUrl = new URL('/auth/login', request.url)
         loginUrl.searchParams.set('redirect_to', request.nextUrl.pathname)
         return NextResponse.redirect(loginUrl)
     }
 
-    // ユーザーが認証されている場合、リクエストを続行します。
-    return NextResponse.next()
+    // 認証APIを呼び出してセッションを検証
+    const responseAPI = await fetch(new URL('/api/auth/verify', request.url), {
+        method: 'POST',
+        headers: {
+            Authorization: `Bearer ${session}`,
+        },
+    });
 
+    // 認証されていない場合はログインページへリダイレクト
+    if (responseAPI.status !== 200) {
+        const loginUrl = new URL('/auth/login', request.url);
+        loginUrl.searchParams.set('redirect_to', request.nextUrl.pathname)
+        return NextResponse.redirect(loginUrl);
+    }
+
+    return NextResponse.next()
 }
+
 // 認証が必要なルートを指定
 export const config = {
     matcher: [
